requirements.txt
# shared
fastapi[standard]
httpx
pydantic-settings
rich
qdrant-client

# embedder
langchain-huggingface
sentence_transformers

# indexer
langchain-text-splitters
langchain-community
pypdf
faster-whisper

# chat
openai
openai-agents
